feat: rewrite from scratch
The underlying Fetch implementation is rewritten from scratch, due to serious issues and lack of
support of the previously used implementation (fetch-h2).

BREAKING CHANGE: The exposed API, options etc have been aligned with node-fetch. Due to the major
changes there are numerous breaking changes both in the API and the options.

fixes #20, #27, #32, #81